refactor(event): 将重试次数相关字段从 int 改为 long 类型
- 修改 OutboxMessage 中 retryCount 和 maxRetries 字段类型为 long
- 调整 OutboxRepository 中对应数据库读写时的类型处理为 long
- 修正 SearchResponse 中 isDegraded 字段命名符合代码规范
- 新增用户画像表 ydsz_agt_user_profile，支持用户偏好与长期画像数据存储
- 新增洞察报告表 ydsz_agt_insight_report，存储 BI 报告及分析结果 JSON
- 实现自动更新时间戳的 PostgreSQL 触发器，替代 MySQL 的 ON UPDATE CURRENT_TIMESTAMP 机制

+++ ydsz-common/ydsz-common-event/src/main/java/com/njydsz/common/event/model/OutboxMessage.java
@@ -61,10 +61,10 @@ public class OutboxMessage {
   private final OutboxStatus status;
 
   /** 重试次数 */
-  private final int retryCount;
+  private final long retryCount;
 
   /** 最大重试次数 */
-  private final int maxRetries;
+  private final long maxRetries;
 
   /** 下次重试时间（指数退避） */
   private final Instant nextRetryAt;
